+++ apps/backend/src/libs/modules/server-application/base-server-application.ts
@@ -193,6 +193,7 @@ class BaseServerApplication implements ServerApplication {
 			preHandler: preHandlers ?? [],
 			schema: {
 				body: validation?.body,
+				query: validation?.query,
 			},
 			url: path,
 		};