+++ apps/frontend/src/libs/modules/store/libs/middlewares/middlewares.ts
@@ -1,1 +1,2 @@
 export { handleError as handleErrorMiddleware } from "./handle-error.middleware.js";
+export { handleUnauthorizedError as handleUnauthorizedErrorMiddleware } from "./handle-unauthorized-error.middleware.js";