+++ apps/frontend/src/pages/access-management/libs/components/components.ts
@@ -3,4 +3,5 @@ export { GroupMenu } from "./group-menu/group-menu.js";
 export { GroupUpdateForm } from "./group-update-form/group-update-form.js";
 export { GroupUsersTable } from "./group-users-table/group-users-table.js";
 export { GroupsTable } from "./groups-table/groups-table.js";
+export { UserMenu } from "./user-menu/user-menu.js";
 export { UsersTable } from "./users-table/users-table.js";