+++ apps/frontend/src/pages/access-management/libs/helpers/get-user-columns/get-user-columns.helper.tsx
@@ -1,9 +1,12 @@
 import { formatDate } from "~/libs/helpers/helpers.js";
 import { type TableColumn } from "~/libs/types/types.js";
 
+import { UserMenu } from "../../components/components.js";
 import { type UserRow } from "../../types/types.js";
 
-const getUserColumns = (): TableColumn<UserRow>[] => {
+const getUserColumns = (actions: {
+	onDelete: (userId: number) => void;
+}): TableColumn<UserRow>[] => {
 	return [
 		{
 			accessorKey: "name",
@@ -13,13 +16,21 @@ const getUserColumns = (): TableColumn<UserRow>[] => {
 		{
 			accessorFn: (user: UserRow): string => user.groups.join(", "),
 			header: "Groups",
-			size: 460,
+			size: 400,
 		},
 		{
 			accessorFn: (user: UserRow): string =>
 				formatDate(new Date(user.createdAt), "d MMM yyyy HH:mm"),
 			header: "Created At",
 			size: 200,
+		},
+		{
+			cell: ({ row: { original: user } }) => (
+				<UserMenu onDelete={actions.onDelete} userId={user.id} />
+			),
+			header: "",
+			id: "menu",
+			size: 60,
 		},
 	];
 };