+++ apps/frontend/src/pages/project/libs/components/setup-analytics-modal/setup-analytics-modal.tsx
@@ -15,6 +15,7 @@ import {
 } from "~/libs/hooks/hooks.js";
 import { actions as projectApiKeyActions } from "~/modules/project-api-keys/project-api-keys.js";
 import { type ProjectGetByIdResponseDto } from "~/modules/projects/projects.js";
+import { actions as scriptActions } from "~/modules/scripts/scripts.js";
 
 import styles from "./styles.module.css";
 
@@ -41,7 +42,9 @@ const SetupAnalyticsModal = ({
 	const isCopyButtonDisabled =
 		!hasProjectApiKey || dataStatus === DataStatus.PENDING;
 
-	const script = useMemo<string>(() => {
+	const pm2StartupScript = "pm2 startup";
+
+	const analyticsScript = useMemo<string>(() => {
 		if (!hasProjectApiKey || !hasAuthenticatedUser) {
 			return "";
 		}
@@ -54,9 +57,10 @@ const SetupAnalyticsModal = ({
 
 	const { control, errors, handleSubmit, handleValueSet } = useAppForm({
 		defaultValues: {
+			analyticsScript,
 			apiKey: project.apiKey ?? "",
+			pm2StartupScript,
 			projectId: project.id,
-			script,
 		},
 	});
 
@@ -69,28 +73,39 @@ const SetupAnalyticsModal = ({
 		[handleSubmit, dispatch],
 	);
 
-	const handleCopyToClipboard = useCallback(
+	const handleCopyApiKeyToClipboard = useCallback(
 		(input: string) => {
 			void dispatch(projectApiKeyActions.copyToClipboard(input));
 		},
 		[dispatch],
 	);
 
-	const handleCopyAPIKeyClick = useCallback(() => {
-		handleCopyToClipboard(project.apiKey as string);
-	}, [handleCopyToClipboard, project]);
+	const handleCopyScriptToClipboard = useCallback(
+		(input: string) => {
+			void dispatch(scriptActions.copyToClipboard(input));
+		},
+		[dispatch],
+	);
 
-	const handleCopyScriptClick = useCallback(() => {
-		handleCopyToClipboard(script);
-	}, [handleCopyToClipboard, script]);
+	const handleCopyAPIKeyClick = useCallback(() => {
+		handleCopyApiKeyToClipboard(project.apiKey as string);
+	}, [handleCopyApiKeyToClipboard, project]);
+
+	const handleCopyAnalyticsScriptClick = useCallback(() => {
+		handleCopyScriptToClipboard(analyticsScript);
+	}, [handleCopyScriptToClipboard, analyticsScript]);
+
+	const handleCopyStartupScriptClick = useCallback(() => {
+		handleCopyScriptToClipboard(pm2StartupScript);
+	}, [handleCopyScriptToClipboard, pm2StartupScript]);
 
 	useEffect(() => {
 		handleValueSet("apiKey", project.apiKey ?? "");
 	}, [handleValueSet, project.apiKey]);
 
 	useEffect(() => {
-		handleValueSet("script", script);
-	}, [handleValueSet, script]);
+		handleValueSet("analyticsScript", analyticsScript);
+	}, [handleValueSet, analyticsScript]);
 
 	return (
 		<Modal isOpened={isOpened} onClose={onClose} title="Setup Analytics">
@@ -143,12 +158,53 @@ const SetupAnalyticsModal = ({
 							<span className={styles["list-item-title"]}>Node.js 20</span>: The
 							script requires Node.js 20 to be installed on your machine.
 						</li>
+						<li>
+							<span className={styles["list-item-title"]}>PM2 5.4</span>: The
+							script requires PM2 5.4 to be installed on your machine.
+						</li>
+						<li>
+							<span className={styles["list-item-title"]}>
+								Unix-based system
+							</span>
+							: The script requires a Unix-based operating system (e.g., Linux
+							or macOS) to run properly.
+						</li>
 					</ul>
 				</div>
 
 				<div>
 					<span className={styles["subtitle"]}>Installation Steps</span>
 					<ol className={styles["text"]}>
+						<li className={styles["list-item"]}>
+							<span className={styles["list-item-title"]}>
+								Execute the configuration script.
+							</span>
+							<p className={styles["list-item-text"]}>
+								Open your terminal or console, copy the following script, paste
+								and run it.
+							</p>
+							<Input
+								control={control}
+								errors={errors}
+								isLabelHidden
+								isReadOnly
+								label="pm2StartupScript"
+								name="pm2StartupScript"
+								placeholder="Need to generate API key"
+								rightIcon={
+									<IconButton
+										iconName="clipboard"
+										isDisabled={isCopyButtonDisabled}
+										label="Copy script"
+										onClick={handleCopyStartupScriptClick}
+									/>
+								}
+							/>
+							<span className={styles["list-item-text"]}>
+								Then, modify the command output from this step with your
+								system&apos;s values and execute it in terminal or console.
+							</span>
+						</li>
 						<li className={styles["list-item"]}>
 							<span className={styles["list-item-title"]}>
 								Clone your project repository.
@@ -163,23 +219,23 @@ const SetupAnalyticsModal = ({
 								Prepare the script.
 							</span>
 							<p className={styles["list-item-text"]}>
-								Copy the command below and modify with your local
-								repository&apos;s path:
+								Copy the command below and replace &lt;project-path&gt;
+								placeholder with your local repository&apos;s path:
 							</p>
 							<Input
 								control={control}
 								errors={errors}
 								isLabelHidden
 								isReadOnly
-								label="Script"
-								name="script"
+								label="Analytics script"
+								name="analyticsScript"
 								placeholder="Need to generate API key"
 								rightIcon={
 									<IconButton
 										iconName="clipboard"
 										isDisabled={isCopyButtonDisabled}
 										label="Copy script"
-										onClick={handleCopyScriptClick}
+										onClick={handleCopyAnalyticsScriptClick}
 									/>
 								}
 							/>
@@ -191,7 +247,7 @@ const SetupAnalyticsModal = ({
 							</span>
 							<p className={styles["list-item-text"]}>
 								Open your terminal or console, paste and run the modified
-								script.
+								script. Script will start and be saved to restart on reboot.
 							</p>
 						</li>
 					</ol>